Add files via upload
UMLPI HALAL ECOSYSTEM: A PROJECT PROPOSAL -- The United Muslim Leaders of the Philippines Inc (UMLPI) shall sign an agreement with the GoHalal and the Local Government Units, within which GoHalal is given the opportunity to assist the implementation of the overall initiatives for a Halal Hub in Central Luzon including the development of blueprint, systems and framework for the Halal Hub.

With an objective to position Central Luzon in the global Halal Market, the first phase will see GoHalal provision in Halal Consultation, Advisory and Development to governments, industry stakeholders, service providers and other agencies on numerous requirements, such as: Halal hub development & initiatives; Poultry & meat-based project development; Pharmaceutical industry development; Halal logistics dynamics; Halal e-Commerce development; Market intelligence and analysis; Master plan development; Marketing and branding strategy development; and, Public relations & event management, among others.

It is also intended that GoHalal will be involved in the establishment of a Halal Accreditation Center for the Halal Hub, and the development and design of a Halal Trade and Import-Export Ecosystem which encompasses the Halal Logistics and Supply Chain, Halal Marketplace, Halal Governance, Certification and Compliance System, Halal Verification Rapid Test Kit, and Halal Track and Trace System.

The United Muslim Leaders of the Philippines Inc (UMLPI) is an organization of Muslim community leaders and affiliate traders associations and community groups in the cities and the provinces supporting the government's goal of the Halal Industry’s positioning of Central Luzon as a Muslim friendly tourist Destination.

GoHalal (www.gohalal4all.com) is an integrated Halal eMarketplace that caters for both Business-to-Consumer retailing as well as Business-to-Business transactions. It offers easy access to a secured electronic platform leveraging on the Web portal, mobile services and social media, designed to help the entrepreneurs to gain brand exposure and tap on global business opportunities from parties across around the world, particularly from Halal-savvy consumer markets.