Add files via upload
BULACAN has an estimated Muslim population of 95,000 with its leaders actively engaging in nation building, and particularly, supporting the Muslim-friendly Administration of Governor Daniel Fernando. With its vigorous leadership in the Bulacan Muslim Affairs Consultative Council (BMACC), the Governor is now in the best position to spearhead a model Halal Ecosystem and develop all its local potentials to become the center-stage for stimulating the Philippine Halal Market Potential of US$400 Billion.